Retype taskId to Long and taskOrder to Integer.

Closes #12

See merge request muni-kypo-crp/backend-java/kypo-elasticsearch-documents!12

+++ src/main/java/cz/muni/csirt/kypo/events/adaptive/trainings/AbstractAuditAdaptivePOJO.java
@@ -63,10 +63,10 @@ public abstract class AbstractAuditAdaptivePOJO {
     protected long phaseOrder;
     @ApiModelProperty(value = "Training task ID.", required = true)
     @JsonProperty(value = "task_id", required = true)
-    protected long taskId;
+    protected Long taskId;
     @ApiModelProperty(value = "Training task order.", required = true)
     @JsonProperty(value = "task_order", required = true)
-    protected long taskOrder;
+    protected Integer taskOrder;
     @ApiModelProperty(value = "Id of player in the training run")
     @JsonProperty(value = "user_ref_id")
     protected long userRefId;
